+++ src/rendering/IndexBuffer.cpp
@@ -1,6 +1,6 @@
 #include <rendering/IndexBuffer.h>
 
-IndexBuffer::IndexBuffer(std::vector<unsigned int> indices)
+IndexBuffer::IndexBuffer(std::vector<unsigned short> indices)
 {
 	this->indices = indices;
 	glGenBuffers(1, &this->ID);
@@ -10,7 +10,7 @@ IndexBuffer::IndexBuffer(std::vector<unsigned int> indices)
 void IndexBuffer::init()
 {
 	glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, this->ID);
-	glBufferData(GL_ELEMENT_ARRAY_BUFFER, indices.size() * sizeof(unsigned int), &indices[0], GL_STATIC_DRAW);
+	glBufferData(GL_ELEMENT_ARRAY_BUFFER, indices.size() * sizeof(unsigned short), &indices[0], GL_STATIC_DRAW);
 }
 
 void IndexBuffer::bind()
